feat: Add image background and text overlay to Compose UI
Adds a new composable function, `PractiseOne`, which displays an image as a background with overlaid text. The `androidparty` image is loaded from resources, and the text "Happy Birthday Sam!" and "From Emma" are displayed. The image is set to have 0.5f transparency and `ContentScale.Crop` and text is displayed using `Column` and `Text` composables. The text "Happy Birthday Sam!" is styled with a font size of 100.sp, centered alignment, and a line height of 120.sp. "From Emma" is styled with 32.sp and centered. Also creates `PractiseOnePreview` to display in preview.