feat: Add contentDescription explanation for accessibility
Adds a new file `05_contentDescription.kt` to explain the purpose of `contentDescription` in Android. It clarifies that `contentDescription` is used to provide textual descriptions of images for screen readers, enhancing accessibility. It also notes that `contentDescription` is not intended for use in application logic or callbacks and should be set to null for decorative images.