feat: Implement ImeAction in Jetpack Compose
Adds `ImeActionFunction` composable to demonstrate the use of `ImeAction` in Jetpack Compose. This allows specifying the action to be performed when the user presses the action button on the keyboard. The composable contains an `OutlinedTextField` and demonstrates setting `ImeAction` to `Search`, `Next`, `Send` or `Go`. Also, adds `PreviewImeActionFunction` for preview.